scout/requirements.txt
vdf>=3.4
boto3>=1.34
python-dotenv>=1.0
PyYAML>=6.0
pytest>=8.0
pytest-mock>=3.0
